long long -> uint64_t
Use properly sized types rather than 'this type is at least this big' types.

+++ crawl-ref/source/perlin.cc
@@ -100,11 +100,11 @@ namespace perlin
     const double F4 = (sqrt(5.0)-1.0)/4.0;
     const double G4 = (5.0-sqrt(5.0))/20.0;
 
-    // Use long long so that noise() can work sensibly for
+    // Use uint64_t so that noise() can work sensibly for
     // coordinates from the full range of uint32_t; otherwise scaling,
     // signedness, and skew will give us considerably less than that.
-    static long long fastfloor(const double x) {
-        long long xi = (long long)x;
+    static uint64_t fastfloor(const double x) {
+        uint64_t xi = (uint64_t) x;
         return x<xi ? xi-1 : xi;
     }
 
@@ -124,8 +124,8 @@ namespace perlin
         double n0, n1, n2; // Noise contributions from the three corners
         // Skew the input space to determine which simplex cell we're in
         double s = (xin+yin)*F2; // Hairy factor for 2D
-        long long i = fastfloor(xin+s);
-        long long j = fastfloor(yin+s);
+        uint64_t i = fastfloor(xin+s);
+        uint64_t j = fastfloor(yin+s);
         double t = (i+j)*G2;
         double X0 = i-t; // Unskew the cell origin back to (x,y) space
         double Y0 = j-t;
@@ -178,9 +178,9 @@ namespace perlin
         double n0, n1, n2, n3; // Noise contributions from the four corners
         // Skew the input space to determine which simplex cell we're in
         double s = (xin+yin+zin)*F3; // Very nice and simple skew factor for 3D
-        long long i = fastfloor(xin+s);
-        long long j = fastfloor(yin+s);
-        long long k = fastfloor(zin+s);
+        uint64_t i = fastfloor(xin+s);
+        uint64_t j = fastfloor(yin+s);
+        uint64_t k = fastfloor(zin+s);
         double t = (i+j+k)*G3;
         double X0 = i-t; // Unskew the cell origin back to (x,y,z) space
         double Y0 = j-t;
@@ -261,10 +261,10 @@ namespace perlin
         double n0, n1, n2, n3, n4; // Noise contributions from the five corners
         // Skew the (x,y,z,w) space to determine which cell of 24 simplices we're in
         double s = (x + y + z + w) * F4; // Factor for 4D skewing
-        long long i = fastfloor(x + s);
-        long long j = fastfloor(y + s);
-        long long k = fastfloor(z + s);
-        long long l = fastfloor(w + s);
+        uint64_t i = fastfloor(x + s);
+        uint64_t j = fastfloor(y + s);
+        uint64_t k = fastfloor(z + s);
+        uint64_t l = fastfloor(w + s);
         double t = (i + j + k + l) * G4; // Factor for 4D unskewing
         double X0 = i - t; // Unskew the cell origin back to (x,y,z,w) space
         double Y0 = j - t;
